# url/storage/api-docs/api-docs.json
swagger: "2.0"
info:
  description: "API для swagger.json сокращателя ссылок"
  version: "1.0.0"
  title: "Swagger api shortner"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "192.168.0.63"
basePath: "/api"
...
schemes:
- "http"
paths:
  /showAllLinks:
    get:
      tags:
      - "store"
      summary: "Возвращает все ссылки"
      description: "Возвращает все ссылки"
      operationId: "showAllLinks"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
              url:
                type: "string"
              code:
                type: "string"
              created_at:
                type: "string"
              updated_at:
                type: "string"
              hits:
                type: "integer"
      security:
      - api_key: []
...